make reusable hooks to prevent too much copy
