+++ src/components/testCaseLanding/common/Hooks/UseToast.tsx
@@ -1,0 +1,23 @@
+import React, { useState } from "react";
+
+function UseToast() {
+  const [toastOpen, setToastOpen] = useState<boolean>(false);
+  const [toastMessage, setToastMessage] = useState<string>("");
+  const [toastType, setToastType] = useState<string>("danger");
+  const onToastClose = () => {
+    setToastType("danger");
+    setToastMessage("");
+    setToastOpen(false);
+  };
+
+  return {
+    toastOpen,
+    setToastOpen,
+    toastMessage,
+    setToastMessage,
+    toastType,
+    setToastType,
+    onToastClose,
+  };
+}
+export default UseToast;
