+++ src/components/testCaseLanding/qdm/TestCaseLanding.test.tsx
@@ -48,6 +48,7 @@ jest.mock("@madie/madie-util", () => ({
   useFeatureFlags: jest.fn().mockImplementation(() => ({
     applyDefaults: false,
     importTestCases: false,
+    qdmTestCases: true,
   })),
 }));
